requirements.txt
pytest
pytest-pycodestyle
flask
requests
datetime
testfixtures
pymodm
pymongo[srv]
Pillow
biosppy
matplotlib
scikit-image